support continue train from checkpoint
