+++ Fine_Tune/util/__init__.py
@@ -1,4 +1,4 @@
 from .cnn_train_helper import train_eval as cnn_train_eval, save_model
 from .flickr8k_dataloader import Flickr8kDataset, get_dataloader
 from .train_helper import train_eval
-from .multi_part_model_serialize import load_model_chunks, save_model_chunks
+from .model_serilizer import load_model_chunks, save_model_chunks, load_state_dicts, save_state_dicts
